# backend/src/main/java/com/forge/habit/controller/HabitReminderController.java
package com.forge.habit.controller;

import com.forge.habit.dto.HabitReminderRequest;
import com.forge.habit.dto.HabitReminderResponse;
import com.forge.habit.service.HabitReminderService;
import jakarta.validation.Valid;
import java.util.List;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/** See docs/api/openapi.yaml `/habits/{id}/reminders`. */
@RestController
@RequestMapping("/api/v1/habits/{habitId}/reminders")
public class HabitReminderController {

    private final HabitReminderService reminderService;

    public HabitReminderController(HabitReminderService reminderService) {
        this.reminderService = reminderService;
    }

    @GetMapping
    public List<HabitReminderResponse> list(@PathVariable UUID habitId) {
        return reminderService.list(habitId);
    }

    @PostMapping
    public ResponseEntity<HabitReminderResponse> create(@PathVariable UUID habitId, @Valid @RequestBody HabitReminderRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(reminderService.create(habitId, request));
    }

    @DeleteMapping("/{reminderId}")
    public ResponseEntity<Void> delete(@PathVariable UUID habitId, @PathVariable UUID reminderId) {
        reminderService.delete(habitId, reminderId);
        return ResponseEntity.noContent().build();
    }
}